Daily report: only display non empty result tables to avoid cluttering the UI.
Summary:
This is a useability fix only. Empty tables are only taking space and
this is distracting from the essential overview the daily report page is
supposed to provide. However, this could be considered controversial as
some could prefer to keep the empty tables, thus this commit review
solliciting other users' feedback to check if there is agreement.

Reviewers: kristof.beyls, cmatthews, llvm-commits

Subscribers: mcrosier

Differential Revision: https://reviews.llvm.org/D29991