Update sponsors list and reformat the web page while there.
This attempts cleanup a bit the html code for the page. We also add more space around the logos, as
there are usually guidelines when using company logos regarding the margin to other elements in the page.
Finally, whithin a sponsor category, logos are displayed in rows, as this avoids having an otherwise
tall and skiny page where the sponsors at the bottom have almost no chance to be seen.